:building_construction: Split Utils into Ext files
